Initial tree prototype Arduino script and tree model (fbx)
I had this in my Unity test scene. I'm going to start integrating it into this scene now. At the moment this script is freestanding and doesn't interact with Unity.

The treePrototype model I made for 3D printing, if we want to use it.. Also made an fbx version we can use in Unity if needed.